Add include to deal with local/external image (#3)
The file `local-or-external-image.html` should take one parameter
`target` which is the raw string to be tested for whether it is local or
external. It then "returns" by way of templating the particular path
that we want, whether that is a path that has `/imagas/` prepended or
not.

NOTE: This is a breaking change. Any avatar image, feature image, or
background image which previously used the path `/image/my-pic.jpg` will
have to be changed to drop the `/images/` portion.